Check sign rotation before using cached ProtoRepairs
Fixes an edge case that occurred if a repair sign was precisely on the craft's center of rotation. The issue occurred if a player right-clicked the sign once, rotated the craft, and then right-clicked the sign again in quick succession. This caused the repair to begin as if the craft had not been rotated.