Rotate repair states based on the rotation of the sign (#51)
* Add support for directional blocks to getRotation functions in RotationUtils

* Simplify calculation of schematic sign position

* Rotate repair state based on the rotation of the sign

* Use WorldEdit's BukkitAdapter to determine the rotation of schematic blocks

* Check sign rotation before using cached ProtoRepairs

Fixes an edge case that occurred if a repair sign was precisely on the craft's center of rotation. The issue occurred if a player right-clicked the sign once, rotated the craft, and then right-clicked the sign again in quick succession. This caused the repair to begin as if the craft had not been rotated.

* Expand exceptions to add rotation

* Clean up offset math

---------

Co-authored-by: TylerS1066 <TylerS1066@users.noreply.github.com>